requirements.txt
fastapi~=0.119.0
uvicorn[standard]~=0.30.6
pydantic~=2.12.2
sqlalchemy~=2.0.34
databases[aiosqlite]~=0.9.0
passlib[bcrypt]~=1.7.4
python-jose~=3.3.0
python-multipart~=0.0.9
